Dev : moved TPS fetches to client side
