--- packages/arb-token-bridge-ui/src/hooks/useArbStats.ts
@@ -1,30 +1,0 @@
-import useSWR from 'swr'
-import { getAPIBaseUrl } from '../util'
-import { ArbStats } from '../pages/api/arbstats'
-import { useNetworksAndSigners } from './useNetworksAndSigners'
-
-export const useArbStats = () => {
-  const {
-    l2: {
-      network: { chainID: l2ChainId }
-    }
-  } = useNetworksAndSigners()
-
-  return useSWR(
-    ['arbstats', l2ChainId],
-    async () => {
-      const response = await fetch(
-        `${getAPIBaseUrl()}/api/arbstats?l2ChainId=${l2ChainId}`,
-        {
-          method: 'GET',
-          headers: { 'Content-Type': 'application/json' }
-        }
-      )
-      return (await response.json()).data as ArbStats
-    },
-    {
-      refreshInterval: 30_000,
-      shouldRetryOnError: false
-    }
-  )
-}
